requirements.txt
django>=4.2
django-redis==5.4.0
django-storages==1.14.4
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.0.0

psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7

gunicorn==23.0.0
urllib3==1.26.20
pillow==10.4.0

redis
requests
django-currentuser
django-cors-headers